+++ Human.cc
@@ -9,7 +9,7 @@ bool Human::makeHumanMove(Vec start, Vec end){
     // cout << "Number of Legal Moves for Player " << colour << ' ' << legalMoves.size() << endl;
     // cout << endl; 
 
-    for (int i = 0; i < legalMoves.size(); ++i) {
+    for (size_t i = 0; i < legalMoves.size(); ++i) {
         // cout << "start: " << start << endl;
         // cout << "end: " << end << endl;
         // cout << "start legal move: " << legalMoves[i][0] << endl;
